requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
lxml==4.9.3

# LLM and Embeddings
google-generativeai==0.3.1
langchain==0.1.0
langchain-google-genai==0.0.5
chromadb==0.4.18
sentence-transformers==2.2.2

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Additional Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1

# CORS
fastapi-cors==0.0.6

# Testing and Evaluation
pytest==7.4.3
httpx==0.25.2
